+++ R/fays_generalized_replication.R
@@ -167,22 +167,25 @@
 
 make_fays_gen_rep_factors <- function(
     Sigma,
-    max_replicates = Matrix::rankMatrix(Sigma) + 4,
+    max_replicates = Inf,
     balanced = TRUE
 ) {
 
   n <- nrow(Sigma)
 
   # Calculate spectral decomposition ----
-  eigen_decomposition <- eigen(x = Sigma, symmetric = TRUE)
+  eigen_decomposition <- compute_eigen_decomposition(Sigma)
+  
   Sigma_rank <- Matrix::rankMatrix(Sigma, method = "qr")
 
   # Obtain eigenvectors scaled by square roots of eigenvalues ----
   v <- sapply(X = seq_along(eigen_decomposition$values),
               FUN = function(k) {
-                truncated_eigenvalue <- ifelse(eigen_decomposition$values[k] < 0,
-                                               0, eigen_decomposition$values[k])
-                sqrt(truncated_eigenvalue) * eigen_decomposition$vectors[,k]
+                if (eigen_decomposition$values[k] < 0) {
+                  rep(0, times = nrow(eigen_decomposition$vectors))
+                } else {
+                  sqrt(eigen_decomposition$values[k]) * eigen_decomposition$vectors[,k]
+                }
               })
   v <- v[, seq_len(Sigma_rank), drop=FALSE]
 
@@ -282,9 +285,16 @@ make_fays_gen_rep_factors <- function(
 #'     for balanced sampling designs, proposed by Deville and Tillé (2005).
 #'   \item \strong{"SD1"}: \cr The non-circular successive-differences variance estimator described by Ash (2014),
 #'     sometimes used for variance estimation for systematic sampling.
-#'   \item\strong{"SD2"}:  \cr The circular successive-differences variance estimator described by Ash (2014).
+#'   \item \strong{"SD2"}:  \cr The circular successive-differences variance estimator described by Ash (2014).
 #'     This estimator is the basis of the "successive-differences replication" estimator commonly used
 #'     for variance estimation for systematic sampling.
+#'   \item \strong{"BOSB"}: \cr The kernel-based variance estimator proposed by
+#'     Breidt, Opsomer, and Sanchez-Borrego (2016) for use with systematic samples
+#'     or other finely stratified designs. Uses the Epanechnikov kernel
+#'     with the bandwidth automatically chosen to result in the smallest possible
+#'     nonempty kernel window.
+#'   \item\strong{"Beaumont-Emond"}: \cr The variance estimator of Beaumont and Emond (2022)
+#'     for multistage unequal-probability sampling without replacement.
 #' }
 #' @param aux_var_names (Only used if \code{variance_estimator = "Deville-Tille")}.
 #' A vector of the names of auxiliary variables used in sampling.
@@ -313,7 +323,7 @@ make_fays_gen_rep_factors <- function(
 #' for details of the approximation.
 #' @param compress This reduces the computer memory required to represent the replicate weights and has no
 #' impact on estimates.
-#' @param mse If \code{TRUE} (the default), compute variances from sums of squares around the point estimate from the full-sample weights,
+#' @param mse If \code{TRUE} (the default), compute variances from sums of squares around the point estimate from the full-sample weights.
 #' If \code{FALSE}, compute variances from sums of squares around the mean estimate from the replicate weights.
 #' For Fay's generalized replication method, setting \code{mse = FALSE} can potentially
 #' lead to large underestimates of variance.
@@ -365,6 +375,15 @@ make_fays_gen_rep_factors <- function(
 #' - Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
 #' \strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.
 #' \cr \cr
+#' - Beaumont, J.-F.; Émond, N. (2022).
+#' "\emph{A Bootstrap Variance Estimation Method for Multistage Sampling and Two-Phase Sampling When Poisson Sampling Is Used at the Second Phase}."
+#' \strong{Stats}, \emph{5}: 339–357.
+#' https://doi.org/10.3390/stats5020019
+#' \cr \cr
+#' - Breidt, F. J., Opsomer, J. D., & Sanchez-Borrego, I. (2016). 
+#' "\emph{Nonparametric Variance Estimation Under Fine Stratification: An Alternative to Collapsed Strata}." 
+#' \strong{Journal of the American Statistical Association}, 111(514), 822–833. https://doi.org/10.1080/01621459.2015.1058264
+#' \cr \cr
 #' - Deville, J.‐C., and Tillé, Y. (2005). "\emph{Variance approximation under balanced sampling.}"
 #' \strong{Journal of Statistical Planning and Inference}, 128, 569–591.
 #' \cr \cr
@@ -387,9 +406,8 @@ make_fays_gen_rep_factors <- function(
 #'   data('library_stsys_sample', package = 'svrep')
 #'
 #'   ## First, ensure data are sorted in same order as was used in sampling
-#'   library_stsys_sample <- library_stsys_sample[
-#'     order(library_stsys_sample$SAMPLING_SORT_ORDER),
-#'   ]
+#'   library_stsys_sample <- library_stsys_sample |>
+#'     sort_by(~ SAMPLING_SORT_ORDER)
 #'
 #'   ## Create a survey design object
 #'   design_obj <- svydesign(
@@ -413,7 +431,7 @@ make_fays_gen_rep_factors <- function(
 #' @export
 as_fays_gen_rep_design <- function(design, variance_estimator = NULL,
                                    aux_var_names = NULL,
-                                   max_replicates = 500,
+                                   max_replicates = Inf,
                                    balanced = TRUE,
                                    psd_option = "warn",
                                    mse = TRUE,
@@ -429,7 +447,7 @@ as_fays_gen_rep_design <- function(design, variance_estimator = NULL,
 #' @export
 as_fays_gen_rep_design.twophase2 <- function(design, variance_estimator = NULL,
                                              aux_var_names = NULL,
-                                             max_replicates = 500,
+                                             max_replicates = Inf,
                                              balanced = TRUE,
                                              psd_option = "warn",
                                              mse = getOption("survey.replicates.mse"),
@@ -493,7 +511,7 @@ as_fays_gen_rep_design.twophase2 <- function(design, variance_estimator = NULL,
 #' @export
 as_fays_gen_rep_design.survey.design <- function(design, variance_estimator = NULL,
                                                  aux_var_names = NULL,
-                                                 max_replicates = 500,
+                                                 max_replicates = Inf,
                                                  balanced = TRUE,
                                                  psd_option = 'warn',
                                                  mse = TRUE,
@@ -577,7 +595,7 @@ as_fays_gen_rep_design.survey.design <- function(design, variance_estimator = NU
 #' @export
 as_fays_gen_rep_design.DBIsvydesign <- function(design, variance_estimator = NULL,
                                                 aux_var_names = NULL,
-                                                max_replicates = 500,
+                                                max_replicates = Inf,
                                                 balanced = TRUE,
                                                 psd_option = 'warn',
                                                 mse = TRUE,