+++ R/variance-estimators.R
@@ -91,6 +91,30 @@
 #' of sampling fractions from earlier stages of sampling. For example, at a third stage of sampling,
 #' the variance estimate from a third-stage stratum is multiplied by \eqn{\frac{n_1}{N_1}\frac{n_2}{N_2}},
 #' which is the product of sampling fractions from the first-stage stratum and second-stage stratum.
+#' @section Beaumont-Emond:
+#' The \strong{"Beaumont-Emond"} variance estimator was proposed by Beaumont and Emond (2022),
+#' intended for designs that use fixed-size, unequal-probability random sampling without replacement.
+#' The variance estimator is simply the Horvitz-Thompson
+#' variance estimator with the following approximation for the joint inclusion
+#' probabilities.
+#' \deqn{
+#'   \pi_{kl} \approx \pi_k \pi_l \frac{n - 1}{(n-1) + \sqrt{(1-\pi_k)(1-\pi_l)}}
+#' }
+#' In the case of cluster sampling, this approximation is
+#' applied to the clusters rather than the units within clusters,
+#' with \eqn{n} denoting the number of sampled clusters. and the probabilities \eqn{\pi}
+#' referring to the cluster's sampling probability. For stratified samples,
+#' the joint probability for units \eqn{k} and \eqn{l} in different strata
+#' is simply the product of \eqn{\pi_k} and \eqn{\pi_l}.
+#' 
+#' For multistage samples, this approximation is applied to the clusters at each stage, separately by stratum.
+#' For later stages of sampling, the variance estimate from a stratum is multiplied by the product
+#' of sampling probabilities from earlier stages of sampling. For example, at a third stage of sampling,
+#' the variance estimate from a third-stage stratum is multiplied by \eqn{\pi_1 \times \pi_{(2 | 1)}},
+#' where \eqn{\pi_1} is the sampling probability of the first-stage unit
+#' and \eqn{\pi_{(2|1)}} is the sampling probability of the second-stage unit
+#' within the first-stage unit.
+#' 
 #' @section Deville 1 and Deville 2:
 #' The \strong{"Deville-1"} and \strong{"Deville-2"} variance estimators
 #' are clearly described in Matei and Tillé (2005),
@@ -124,6 +148,47 @@
 #' where \eqn{\pi_1} is the sampling probability of the first-stage unit
 #' and \eqn{\pi_{(2|1)}} is the sampling probability of the second-stage unit
 #' within the first-stage unit.
+#' @section BOSB:
+#' This kernel-based variance estimator was proposed by Breidt, Opsomer, and Sanchez-Borrego (2016),
+#' for use with samples selected using systematic sampling or where only a single
+#' sampling unit is selected from each stratum (sometimes referred to as "fine stratification").
+#' 
+#' Suppose there are \eqn{n} sampled units, and
+#' for each unit \eqn{i} there is a numeric population characteristic \eqn{x_i}
+#' and there is a weighted total \eqn{\hat{Y}_i}, where
+#' \eqn{\hat{Y}_i} is only observed in the selected sample but \eqn{x_i}
+#' is known prior to sampling.
+#' 
+#' The variance estimator has the following form:
+#' 
+#' \deqn{
+#'   \hat{V}_{ker}=\frac{1}{C_d} \sum_{i=1}^n (\hat{Y}_i-\sum_{j=1}^n d_j(i) \hat{Y}_j)^2
+#' }
+#' 
+#' The terms \eqn{d_j(i)} are kernel weights given by
+#' 
+#' \deqn{
+#'   d_j(i)=\frac{K(\frac{x_i-x_j}{h})}{\sum_{j=1}^n K(\frac{x_i-x_j}{h})}
+#' }
+#' 
+#' where \eqn{K(\cdot)} is a symmetric, bounded kernel function
+#' and \eqn{h} is a bandwidth parameter. The normalizing constant \eqn{C_d} 
+#' is computed as:
+#' 
+#' \deqn{
+#'   C_d=\frac{1}{n} \sum_{i=1}^n(1-2 d_i(i)+\sum_{j=1}^H d_j^2(i))
+#' }
+#' 
+#' For most functions in the 'svrep' package, the kernel function
+#' is the Epanechnikov kernel and the bandwidth is automatically selected
+#' to yield the smallest possible nonempty kernel window, as was recommended
+#' by Breidt, Opsomer, and Sanchez-Borrego (2016). That's the case for
+#' the functions \code{as_fays_gen_rep_design()}, \code{as_gen_boot_design()},
+#' \code{make_quad_form_matrix()}, etc. However, users can construct the quadratic
+#' form matrix of this variance estimator using a different kernel and a different bandwidth
+#' by directly working with the function \code{make_kernel_var_matrix()}.
+#' 
+#' 
 #' @section Deville-Tillé:
 #' See Section 6.8 of Tillé (2020) for more detail on this estimator,
 #' including an explanation of its quadratic form.
@@ -144,9 +209,17 @@
 #' @references
 #' Ash, S. (2014). "\emph{Using successive difference replication for estimating variances}."
 #' \strong{Survey Methodology}, Statistics Canada, 40(1), 47–59.
+#' 
+#' Beaumont, J.-F.; Émond, N. (2022). "\emph{A Bootstrap Variance Estimation Method for Multistage Sampling and Two-Phase Sampling When Poisson Sampling Is Used at the Second Phase}."
+#' \strong{Stats}, \emph{5}: 339–357.
+#' https://doi.org/10.3390/stats5020019
 #'
 #' Bellhouse, D.R. (1985). "\emph{Computing Methods for Variance Estimation in Complex Surveys}."
 #' \strong{Journal of Official Statistics}, Vol.1, No.3.
+#' 
+#' Breidt, F. J., Opsomer, J. D., & Sanchez-Borrego, I. (2016). 
+#' "\emph{Nonparametric Variance Estimation Under Fine Stratification: An Alternative to Collapsed Strata}." 
+#' \strong{Journal of the American Statistical Association}, 111(514), 822–833. https://doi.org/10.1080/01621459.2015.1058264
 #'
 #' Deville, J.‐C., and Tillé, Y. (2005). "\emph{Variance approximation under balanced sampling.}"
 #' \strong{Journal of Statistical Planning and Inference}, 128, 569–591.